feat: add `codeToTokens` API, rename `codeToThemedTokens` to `codeToTokensBase`
