+++ packages/cli/src/code-to-ansi.ts
@@ -1,12 +1,12 @@
 import c from 'chalk'
 import type { BundledLanguage, BundledTheme } from 'shiki'
-import { FontStyle, codeToThemedTokens, getSingletonHighlighter } from 'shiki'
+import { FontStyle, codeToTokensBase, getSingletonHighlighter } from 'shiki'
 import { hexApplyAlpha } from './colors'
 
 export async function codeToANSI(code: string, lang: BundledLanguage, theme: BundledTheme): Promise<string> {
   let output = ''
 
-  const lines = await codeToThemedTokens(code, {
+  const lines = await codeToTokensBase(code, {
     lang,
     theme,
   })
