+++ packages/core/src/code-to-tokens.ts
@@ -1,200 +1,119 @@
-/* ---------------------------------------------------------
- * Copyright (C) Microsoft Corporation. All rights reserved.
- *-------------------------------------------------------- */
-import type { IGrammar } from './textmate'
-import { INITIAL } from './textmate'
-import type { CodeToThemedTokensOptions, FontStyle, ShikiInternal, ThemeRegistrationResolved, ThemedToken, ThemedTokenScopeExplanation, TokenizeWithThemeOptions } from './types'
-import { StackElementMetadata } from './stack-element-metadata'
-import { applyColorReplacements, isNoneTheme, isPlainLang, splitLines } from './utils'
-import { tokenizeAnsiWithTheme } from './code-to-tokens-ansi'
+import { codeToTokensBase } from './code-to-tokens-base'
+import { codeToTokensWithThemes } from './code-to-tokens-themes'
+import type { CodeToTokensOptions, ShikiInternal, ThemedToken, ThemedTokenWithVariants, TokensResult } from './types'
+import { getTokenStyleObject, stringifyTokenStyle } from './utils'
 
-export function codeToThemedTokens(
+/**
+ * High-level code-to-tokens API.
+ *
+ * It will use `codeToTokensWithThemes` or `codeToTokensBase` based on the options.
+ */
+export function codeToTokens(
   internal: ShikiInternal,
   code: string,
-  options: CodeToThemedTokensOptions = {},
-): ThemedToken[][] {
-  const {
-    lang = 'text',
-    theme: themeName = internal.getLoadedThemes()[0],
-  } = options
+  options: CodeToTokensOptions,
+): TokensResult {
+  let bg: string
+  let fg: string
+  let tokens: ThemedToken[][]
+  let themeName: string
+  let rootStyle: string | undefined
 
-  if (isPlainLang(lang) || isNoneTheme(themeName))
-    return splitLines(code).map(line => [{ content: line[0], offset: line[1] }])
+  if ('themes' in options) {
+    const {
+      defaultColor = 'light',
+      cssVariablePrefix = '--shiki-',
+    } = options
 
-  const { theme, colorMap } = internal.setTheme(themeName)
+    const themes = Object.entries(options.themes)
+      .filter(i => i[1])
+      .map(i => ({ color: i[0], theme: i[1]! }))
+      .sort((a, b) => a.color === defaultColor ? -1 : b.color === defaultColor ? 1 : 0)
 
-  if (lang === 'ansi')
-    return tokenizeAnsiWithTheme(theme, code, options)
+    if (themes.length === 0)
+      throw new Error('[shiki] `themes` option must not be empty')
 
-  const _grammar = internal.getLangGrammar(lang)
-  return tokenizeWithTheme(code, _grammar, theme, colorMap, options)
-}
+    const themeTokens = codeToTokensWithThemes(
+      internal,
+      code,
+      options,
+    )
 
-export function tokenizeWithTheme(
-  code: string,
-  grammar: IGrammar,
-  theme: ThemeRegistrationResolved,
-  colorMap: string[],
-  options: TokenizeWithThemeOptions,
-): ThemedToken[][] {
-  const colorReplacements = {
-    ...theme.colorReplacements,
-    ...options?.colorReplacements,
+    if (defaultColor && !themes.find(t => t.color === defaultColor))
+      throw new Error(`[shiki] \`themes\` option must contain the defaultColor key \`${defaultColor}\``)
+
+    const themeRegs = themes.map(t => internal.getTheme(t.theme))
+    const themesOrder = themes.map(t => t.color)
+    tokens = themeTokens
+      .map(line => line.map(token => mergeToken(token, themesOrder, cssVariablePrefix, defaultColor)))
+
+    fg = themes.map((t, idx) => (idx === 0 && defaultColor ? '' : `${cssVariablePrefix + t.color}:`) + (themeRegs[idx].fg || 'inherit')).join(';')
+    bg = themes.map((t, idx) => (idx === 0 && defaultColor ? '' : `${cssVariablePrefix + t.color}-bg:`) + (themeRegs[idx].bg || 'inherit')).join(';')
+    themeName = `shiki-themes ${themeRegs.map(t => t.name).join(' ')}`
+    rootStyle = defaultColor ? undefined : [fg, bg].join(';')
+  }
+  else if ('theme' in options) {
+    tokens = codeToTokensBase(
+      internal,
+      code,
+      options,
+    )
+
+    const _theme = internal.getTheme(options.theme)
+    bg = _theme.bg
+    fg = _theme.fg
+    themeName = _theme.name
+  }
+  else {
+    throw new Error('[shiki] Invalid options, either `theme` or `themes` must be provided')
   }
 
-  const lines = splitLines(code)
+  return {
+    tokens,
+    fg,
+    bg,
+    themeName,
+    rootStyle,
+  }
+}
 
-  let ruleStack = INITIAL
-  let actual: ThemedToken[] = []
-  const final: ThemedToken[][] = []
+function mergeToken(
+  merged: ThemedTokenWithVariants,
+  variantsOrder: string[],
+  cssVariablePrefix: string,
+  defaultColor: string | boolean,
+) {
+  const token: ThemedToken = {
+    content: merged.content,
+    explanation: merged.explanation,
+    offset: merged.offset,
+  }
 
-  for (let i = 0, len = lines.length; i < len; i++) {
-    const [line, lineOffset] = lines[i]
-    if (line === '') {
-      actual = []
-      final.push([])
-      continue
-    }
+  const styles = variantsOrder.map(t => getTokenStyleObject(merged.variants[t]))
 
-    let resultWithScopes
-    let tokensWithScopes
-    let tokensWithScopesIndex
+  // Get all style keys, for themes that missing some style, we put `inherit` to override as needed
+  const styleKeys = new Set(styles.flatMap(t => Object.keys(t)))
+  const mergedStyles = styles.reduce((acc, cur, idx) => {
+    for (const key of styleKeys) {
+      const value = cur[key] || 'inherit'
 
-    if (options.includeExplanation) {
-      resultWithScopes = grammar.tokenizeLine(line, ruleStack)
-      tokensWithScopes = resultWithScopes.tokens
-      tokensWithScopesIndex = 0
-    }
-
-    const result = grammar.tokenizeLine2(line, ruleStack)
-
-    const tokensLength = result.tokens.length / 2
-    for (let j = 0; j < tokensLength; j++) {
-      const startIndex = result.tokens[2 * j]
-      const nextStartIndex = j + 1 < tokensLength ? result.tokens[2 * j + 2] : line.length
-      if (startIndex === nextStartIndex)
-        continue
-
-      const metadata = result.tokens[2 * j + 1]
-      const foreground = StackElementMetadata.getForeground(metadata)
-      const foregroundColor = applyColorReplacements(colorMap[foreground], colorReplacements)
-      const fontStyle: FontStyle = StackElementMetadata.getFontStyle(metadata)
-
-      const token: ThemedToken = {
-        content: line.substring(startIndex, nextStartIndex),
-        offset: lineOffset + startIndex,
-        color: foregroundColor,
-        fontStyle,
+      if (idx === 0 && defaultColor) {
+        acc[key] = value
       }
-
-      if (options.includeExplanation) {
-        token.explanation = []
-        let offset = 0
-        while (startIndex + offset < nextStartIndex) {
-          const tokenWithScopes = tokensWithScopes![tokensWithScopesIndex!]
-
-          const tokenWithScopesText = line.substring(
-            tokenWithScopes.startIndex,
-            tokenWithScopes.endIndex,
-          )
-          offset += tokenWithScopesText.length
-          token.explanation.push({
-            content: tokenWithScopesText,
-            scopes: explainThemeScopes(theme, tokenWithScopes.scopes),
-          })
-
-          tokensWithScopesIndex! += 1
-        }
-      }
-
-      actual.push(token)
-    }
-    final.push(actual)
-    actual = []
-    ruleStack = result.ruleStack
-  }
-
-  return final
-}
-
-function explainThemeScopes(
-  theme: ThemeRegistrationResolved,
-  scopes: string[],
-): ThemedTokenScopeExplanation[] {
-  const result: ThemedTokenScopeExplanation[] = []
-  for (let i = 0, len = scopes.length; i < len; i++) {
-    const parentScopes = scopes.slice(0, i)
-    const scope = scopes[i]
-    result[i] = {
-      scopeName: scope,
-      themeMatches: explainThemeScope(theme, scope, parentScopes),
-    }
-  }
-  return result
-}
-
-function matchesOne(selector: string, scope: string): boolean {
-  const selectorPrefix = `${selector}.`
-  if (selector === scope || scope.substring(0, selectorPrefix.length) === selectorPrefix)
-    return true
-
-  return false
-}
-
-function matches(
-  selector: string,
-  selectorParentScopes: string[],
-  scope: string,
-  parentScopes: string[],
-): boolean {
-  if (!matchesOne(selector, scope))
-    return false
-
-  let selectorParentIndex = selectorParentScopes.length - 1
-  let parentIndex = parentScopes.length - 1
-  while (selectorParentIndex >= 0 && parentIndex >= 0) {
-    if (matchesOne(selectorParentScopes[selectorParentIndex], parentScopes[parentIndex]))
-      selectorParentIndex -= 1
-    parentIndex -= 1
-  }
-
-  if (selectorParentIndex === -1)
-    return true
-
-  return false
-}
-
-function explainThemeScope(
-  theme: ThemeRegistrationResolved,
-  scope: string,
-  parentScopes: string[],
-): any[] {
-  const result: any[] = []
-  let resultLen = 0
-  for (let i = 0, len = theme.settings.length; i < len; i++) {
-    const setting = theme.settings[i]
-    let selectors: string[]
-    if (typeof setting.scope === 'string')
-      selectors = setting.scope.split(/,/).map(scope => scope.trim())
-    else if (Array.isArray(setting.scope))
-      selectors = setting.scope
-    else
-      continue
-
-    for (let j = 0, lenJ = selectors.length; j < lenJ; j++) {
-      const rawSelector = selectors[j]
-      const rawSelectorPieces = rawSelector.split(/ /)
-
-      const selector = rawSelectorPieces[rawSelectorPieces.length - 1]
-      const selectorParentScopes = rawSelectorPieces.slice(0, rawSelectorPieces.length - 1)
-
-      if (matches(selector, selectorParentScopes, scope, parentScopes)) {
-        // match!
-        result[resultLen++] = setting
-        // break the loop
-        j = lenJ
+      else {
+        const keyName = key === 'color' ? '' : key === 'background-color' ? '-bg' : `-${key}`
+        const varKey = cssVariablePrefix + variantsOrder[idx] + (key === 'color' ? '' : keyName)
+        if (acc[key])
+          acc[key] += `;${varKey}:${value}`
+        else
+          acc[key] = `${varKey}:${value}`
       }
     }
-  }
-  return result
+    return acc
+  }, {} as Record<string, string>)
+
+  token.htmlStyle = defaultColor
+    ? stringifyTokenStyle(mergedStyles)
+    : Object.values(mergedStyles).join(';')
+  return token
 }
