+++ packages/core/src/index.ts
@@ -5,7 +5,8 @@ export * from './types'
 
 export { loadWasm } from './oniguruma'
 export { getShikiInternal, setDefaultWasmLoader } from './internal'
-export { codeToThemedTokens, tokenizeWithTheme } from './code-to-tokens'
+export { codeToTokensBase, tokenizeWithTheme } from './code-to-tokens-base'
+export { codeToTokens } from './code-to-tokens'
 export { tokenizeAnsiWithTheme } from './code-to-tokens-ansi'
 export { codeToHast, tokensToHast } from './code-to-hast'
 export { codeToHtml, hastToHtml } from './code-to-html'
