+++ packages/shiki/src/bundle-web.ts
@@ -35,7 +35,8 @@ export const getHighlighter = /* @__PURE__ */ createdBundledHighlighter<
 export const {
   codeToHtml,
   codeToHast,
-  codeToThemedTokens,
+  codeToTokensBase,
+  codeToTokens,
   codeToTokensWithThemes,
   getSingletonHighlighter,
 } = /* @__PURE__ */ createSingletonShorthands<
